// template/vue3-admin/src/utils/http/types.d.ts
import { Method, AxiosError, AxiosResponse, AxiosRequestConfig } from 'axios'

export type resultType = {
  accessToken?: string
}

export type RequestMethods = Extract<
  Method,
  'get' | 'post' | 'put' | 'delete' | 'patch' | 'option' | 'head'
>

export interface PureHttpError extends AxiosError<any> {
  isCancelRequest?: boolean
}

export interface PureHttpResoponse extends AxiosResponse {
  config: PureHttpRequestConfig
}

export interface PureHttpRequestConfig extends AxiosRequestConfig {
  beforeRequestCallback?: (request: PureHttpRequestConfig) => void
  beforeResponseCallback?: (response: PureHttpResoponse) => void
  isRegister?: boolean
  needNProcess?: boolean
}

export default class PureHttp {
  request<T>(
    method: RequestMethods,
    url: string,
    param?: AxiosRequestConfig,
    axiosConfig?: PureHttpRequestConfig
  ): Promise<T>
  post<T, P>(url: string, params?: T, config?: PureHttpRequestConfig): Promise<P>
  get<T, P>(url: string, params?: T, config?: PureHttpRequestConfig): Promise<P>
}